requirements.txt
absl-py==2.1.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
black==24.4.2
bleach==6.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.8
decorator==5.1.1
defusedxml==0.7.1
dm-tree==0.1.8
executing==2.1.0
fastjsonschema==2.20.0
flatbuffers==24.3.25
fonttools==4.55.0
fqdn==1.5.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.67.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.6
httpx==0.27.2
humanfriendly==10.0
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.4
json5==0.9.28
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==3.6.0
kiwisolver==1.4.7
libclang==18.1.1
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
mpmath==1.3.0
msgpack==1.0.8
mypy==1.10.1
mypy-extensions==1.0.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
nodeenv==1.9.1
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
onnx==1.17.0
onnxruntime==1.18.0
opt_einsum==3.4.0
optree==0.13.1
overrides==7.7.0
packaging==24.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.2.2
prometheus_client==0.21.0
prompt_toolkit==3.0.48
protobuf==3.20.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-spy==0.3.14
pycparser==2.22
pygame-ce==2.4.1
Pygments==2.18.0
pyparsing==3.2.0
pyright==1.1.370
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.21.0
Send2Trash==1.8.3
setuptools==75.5.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
sympy==1.12.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-probability==0.25.0
termcolor==2.5.0
terminado==0.18.1
tf2onnx==1.16.1
tf_keras==2.18.0
tinycss2==1.4.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
types-python-dateutil==2.9.0.20241003
types-tqdm==4.66.0.20240417
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==2.2.3
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.0
wrapt==1.16.0
